Change GetTimeSinceStart to double
In the past (13+ years ago), this was float instead of double to maintain compatibility with 32 bit hardware. That is not a concern with anything ITGMania will run on, since ITGMania is 64-bit only, and it's not going to cause noticeable performance issues for any ITGMania users by having this be double. On the other hand, the time precision benefits by having this be a double would be appreciated. Therefore, I am making GetTimeSinceStart a double instead of a float.