Give ThreadedBuffer a handle to change priority
Improve ThreadedBuffer:

BufferingThread(): Calculate unchanging values outside of the loop for better performance

Include windows.h if the OS is Windows so that the sound buffer can be set to high priority

Lines 47 and 48 in .cpp whitespace changes are needed to fix setting the buffer name

---

Hardcode RageSoundReader::ERROR in RageSound: This has to coexist with the other changes in this PR so it will build correctly. We can't use the constant ERROR here as long as windows.h is being indirectly pulled in.