Revert "Fix interaction between global offset and music rate."
This reverts commit 343bbd0e5638bb6f4e8d741473a8fd1a92f2838d.

Global offset is intended to compensate for input and audio latency in the PC
used for playing. This latency stays constant regardless of music rate, so
it's incorrect to multiply it. The original commit states that we need the
multiply to arrive at the same position in the sound file regardless of music
rate, but in fact we want the position to be different for different rates in
order to compensate for the latency correctly.