Calculate global offset in beats
The theory here is that converting the global offset to beats, instead of mixing beats and seconds, should keep the sync tighter when rate modding a song.

The if-else statements inside TimingData::GetElapsedTimeFromBeat(float fBeat) and TimingData::GetBeatAndBPSFromElapsedTime(GetBeatArgs& args) only execute the new code if m_fMusicRate is anything other than 1. They are necessary because stuff will break theme side if it's not done like this. Without the if-else statements, the NPS display and density graph will be incorrect (at least in Simply Love).

phantom10111 is listed as a co-author for their work on #300

Co-Authored-By: phantom10111 <3482004+phantom10111@users.noreply.github.com>