Fix WaveOut buffer underrun
You can give WaveOut as many buffers as you want, but if they are each too large, then we won't fill data as fast as we'd hope to, and that can lead to underrun. So a very simple way to fix this was to increase the number of buffers. When the value for num_chunks is increased, each buffer will be smaller, because chunksize will become smaller as the value of num_chunks increases. The constants were moved from RageSoundDriver_WaveOut.cpp to RageSoundDriver_WaveOut.h to avoid hard-coding the value of num_chunks into the header file.