Create StdString.cpp
This is a refactoring of StdString.h to optimize it for the ITGmania engine. This greatly mitigates the performance shortcomings of RStrings, as well as paving the way to make RStrings unicode compatible.

This is very significant for the engine in the huge performance increase gained here. It was a boost of several hundred FPS on a Ryzen laptop with integrated graphics.

Anything which made sense to migrate to a cpp file has been moved there.

An enhancement is the inclusion of `wchar_t` variants for the `sstoupper` and `sstolower` functions.

I've also marked all functions as noexcept for the sake of allowing the compiler to further optimize, as they don't throw errors.

Anything unused in any way by the game was removed from StdString.h, mostly a lot of unneeded includes.

The original file had way too many comments making it was very hard to get thru so I removed them for the sake of being able to read the code that's here.

As for working Unicode support into this header, but we would need to decide how to handle MakeUpper and MakeLower functions. Besides that, a few functions cast to char* and would need to be adjusted to make Unicode in RString possible.

<cstdarg> include is moved out of StdString into files that need it for va_start / va_end.