Remove skip frames feature in ffmpeg code.
This is a bit too conservative, the old code handled displaying these
types of frames with non-fatal errors--so there's not much reason we
should shy away from taking advantage of it. I tested this with the same
files spurred the addition of the skipping code, and saw no change in
performance.