Bring in std::vector to RageSoundMixBuffer
Now that underlying issues in RageSoundMixBuffer have been resolved, we can use some form of automatic memory management here to replace the malloc/alloc/free calls. This is an implementation using std::vector which ensures that the extended buffer size is a multiple of 1024.

This implementation is safe, fast and functionally identical to the original.

Some functions have been moved from RageSoundMixBuffer.cpp to RageSoundMixBuffer.h as inline functions.

All functions were marked noexcept.

The int_fast64_t data type is used so that the compiler can automatically pick the fastest data type for the platform in question. It's imperative this function completes as fast as possible, so it's better to not find ourselves constrained by the data type used here.