Remove TimingSegment macros from TimingData.h
Un-does a complicated macro. The reasons are that this allows the compiler to better optimize than it would as a macro, and many errors were found to be hiding behind that macro.

Because this replaces the macro with a large section of visually similar code handling different timing segments pointers, a comment has been added explaining the purpose of the functions.