Minimal BitmapText::AddAttribute
Rewrite of commit 208494b

This function indirectly affects the amount of latency before a sound is played. For this reason it has been rewritten with a focus on minimalism while accomplishing the original goal of correcting overlapping attributes. This implementation may not handle all edge cases perfectly, but will handle most cases correctly. This method calculates the starting point for potential overlaps as well as the point of the first element greater than the end position of the new attribute, and calls erase to remove all elements in that range. The new attribute is then inserted into the map.