Remove the SCALE macro
Scales x so that l1 corresponds to l2 and h1 corresponds to h2.
This does not modify x, so it MUST assign the result to something! Do the multiply before the divide to that integer scales have more precision.
One such example: SCALE(x, 0, 1, L, H); interpolate between L and H.