Scale the global offset more accurately when rate modding
The goal of this change is to provide a more accurate offset when using extreme rate mods.  Currently, the farther the rate mod gets from the default value of 1.0, a more drastic offset appears. This aims to mitigate that issue by using the median of the unscaled and scaled global offsets to minimize the offset calculation error.