Streamline message processing logic in LoadingWindow_Win32::Paint()
A potential bottleneck in this code is the window behavior when it regains focus after losing it for a while, where it will briefly max out the CPU trying to process all the missed messages from when it didn't have focus. This commit changes that logic to always process the current message while in focus, and skip everything else if the window does not have focus.