Don't log songwheel audio buffer underrun errors.
Currently,  it's not very useful to diagnose audio buffer issues based off user log.txt's, since the log will fill up with this error every time the user is selecting songs in the song wheel.

This would change RageSoundPosMap to include ScreenManager so the new function GetTopScreenName can be called. We then check if the current screen name is ScreenSelectMusic. If it is, skip the log message.  If it isn't, log the error.

This will make it so that the buffer underrun error is only logged when it needs to be.