Prevent a crash when re-scanning an empty folder in SongManager::LoadSongDir
An empty folder (a folder without audio or simfile) can trigger this crash, because it might attempt to remove a group from the group mapping which was never added in the first place.  Adding a boolean (groupAdded) to track whether or not we need to perform that removal prevents this crash.