Improve reliability of SaveScreenshot
It's possible for a screenshot to occasionally go missing.  This seems to usually be due to one of two reasons:

 - User pushed the screenshot button repeatedly too fast
 - Something went wrong with the path construction

In my testing, NameSuffix is empty almost 100% of the time, and it's more likely that a screenshot fails to save if you explicitly add NameSuffix.  So now we only add NameSuffix if it's not empty.

I've also tested how fast the screenshot button can be pushed for a screenshot to successfully save. It seems like just about 0.5 seconds is the threshold. Any faster and it won't save a screenshot, but will play a success sound. This function makes an asynchronous call to either SavePNG or SaveJPG and does not wait for its response to confirm the screenshot was saved. To keep things simple, I've simply had the function play the invalid sound if we've measured less than 0.5 seconds since the last SaveScreenshot call.

Lastly, really a no-op minor change, but removed a useless boolean since we can simply check if the DISPLAY->SaveScreenshot call failed and use that to return RString.  The successful screenshot sound is also now part of an else clause, to make the logic flow a bit more structured.