Fix for video banner related crashes
By loading a blank graphic BEFORE loading the video,  we should be able to avoid crashes caused by a missing image cache for video files.  It will still generate a logged warning when ImageCache::LoadCachedImage notices there isn't a cached banner, but we'll have loaded the blank banner by then. This is VERY hacky / definitely a band-aid, but should prevent a nasty crash in the songwheel until a more "proper" fix can be implemented.