Fix global offset applied at song loading
Song::ReCalculateStepStatsAndLastSecond used the current value of the
global offset when setting the first second and last second of the song
during song loading. On subsequent launches, the values are loaded from
the song cache. The values are not updated when the global offset is
changed, so they will be wrong if the offset is modified afterwards.

The first second and last second only affect the delay before the music
starts playing and before it stops, so it isn't normally noticeable.

Also changes ScreenEdit to not add the global offset when setting the
last second hint from the area menu.

Also changes song loading to not add the global offset when setting the
default sample start second.